Rename dtask to dtask_srv, add basic dtask module
Make dtask:start() and dtask:stop() operate on an
application level, with dtask_srv the main gen_server.

However, dtask_srv still has the main callbacks,
which is something we want dtask to have (call, cast,
and register). Wasn't sure how to move those conforming
to the original intent in regards to testing and dyalizer.