Switch the gen_servers to gen_leader
Modify the dtask_src and dtask_timer gen_servers to use
gen_leader. They still only work if you treat one of the nodes as the
master and register all the calling nodes with it. To use at this
state you would have to use rpc:call to register and execute against
whichever node you deemed the master.